requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-environ==0.11.2
python-dotenv==1.0.1
django-mssql-backend==2.8.1
django-widget-tweaks==1.5.0
idna==3.10
mssql-django==1.6
pillow==11.3.0
psutil==7.1.0
pyodbc==5.2.0
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
watchdog==6.0.0
whitenoise==6.11.0
